Bump up version in documents, upgrade deployment script

+++ deploy-private.sh
@@ -1,9 +1,17 @@
 #!/bin/bash
 
-git stash
-mvn release:update-versions -DautoVersionSubmodules=true 
-mvn clean deploy -P assembly
-git reset --hard
-git stash apply && git stash drop
+if [[ -n $1 && -n $2 ]];
+then 
+  git stash
+  mvn release:update-versions -DautoVersionSubmodules=true 
+  mvn clean deploy -P assembly
+  git reset --hard
+  find . -type f -exec sed -i 's/$1/$2/g' {} \;
+else
+  echo "Missing version arguments. Usage: $0 [old-version] [new-version]"
+  exit 1
+fi
+
+exit 0
